+++ src/others/PathPreviewer.tsx
@@ -1,52 +1,40 @@
-import { selectDirs } from '../server/appHelper';
-import DirSource from '../helper/DirSource';
-import { useState } from 'react';
-import { useDSEvents } from '../helper/dirSourceHelpers';
+import {
+    copyToClipboard, openExplorer,
+} from '../server/appHelper';
+import appProvider from '../server/appProvider';
+import { showAppContextMenu } from './AppContextMenu';
 
-export default function PathPreviewer({ dirSource }: {
-    dirSource: DirSource,
-    prefix: string
-}) {
-    const [text, setText] = useState(dirSource.dirPath);
-    const [isDirPathValid, setIsDirPathValid] = useState(
-        dirSource.isDirPathValid);
-    useDSEvents(['path'], dirSource, () => {
-        setText(dirSource.dirPath);
-        setIsDirPathValid(dirSource.isDirPathValid);
-    });
-    const applyNewText = (newText: string) => {
-        setText(newText);
-        dirSource.dirPath = newText;
+// TODO: check direction rtl error with /*
+const cleanPath = (path: string) => {
+    if (path && path[0] === '/') {
+        path = path.substring(1);
+    }
+    return path;
+};
+
+export function pathPreviewer(dirPath: string) {
+    const onContextMenu = (event: any) => {
+        showAppContextMenu(event as any, [
+            {
+                title: 'Copy to Clipboard',
+                onClick: () => {
+                    copyToClipboard(dirPath);
+                },
+            },
+            {
+                title: `Reveal in ${appProvider.systemUtils.isMac ?
+                    'Finder' : 'File Explorer'}`,
+                onClick: () => {
+                    openExplorer(dirPath);
+                },
+            },
+        ]);
     };
-    const dirValidCN = isDirPathValid ? 'is-valid' : (
-        isDirPathValid === null ? '' : 'is-invalid');
     return (
-        <div className='input-group mb-3'>
-            {dirSource.dirPath &&
-                <button className='btn btn-secondary'
-                    type='button'
-                    onClick={() => {
-                        return dirSource.fireReloadEvent();
-                    }}>
-                    <i className='bi bi-arrow-clockwise' />
-                </button>
-            }
-            <input type='text'
-                className={`form-control ${dirValidCN}`}
-                value={text}
-                onChange={(event) => {
-                    applyNewText(event.target.value);
-                }} />
-            <button className='btn btn-secondary'
-                type='button'
-                onClick={() => {
-                    const dirs = selectDirs();
-                    if (dirs.length) {
-                        applyNewText(dirs[0]);
-                    }
-                }}>
-                <i className='bi bi-folder2-open' />
-            </button>
+        <div className='ellipsis-left border-white-round px-1 flex-fill'
+            onContextMenu={onContextMenu}
+            title={cleanPath(dirPath)}>
+            {cleanPath(dirPath)}
         </div>
     );
 }
