+++ src/others/PathSelector.tsx
@@ -1,19 +1,14 @@
 import './PathSelector.scss';
 
-import {
-    copyToClipboard,
-    openExplorer,
-} from '../server/appHelper';
-import { showAppContextMenu } from './AppContextMenu';
 import { useStateSettingBoolean } from '../helper/settingHelper';
 import DirSource from '../helper/DirSource';
 import React from 'react';
 import AppSuspense from './AppSuspense';
-import appProvider from '../server/appProvider';
 import { useDSEvents } from '../helper/dirSourceHelpers';
+import { pathPreviewer } from './PathPreviewer';
 
 const PathPreviewer = React.lazy(() => {
-    return import('./PathPreviewer');
+    return import('./PathEditor');
 });
 
 export default function PathSelector({
@@ -26,27 +21,8 @@ export default function PathSelector({
         `${prefix}-selector-opened`, false);
     const dirPath = dirSource.dirPath;
     const isShowingEditor = !dirPath || showing;
-
-    const onContextMenu = (event: any) => {
-        showAppContextMenu(event as any, [
-            {
-                title: 'Copy to Clipboard',
-                onClick: () => {
-                    copyToClipboard(dirPath);
-                },
-            },
-            {
-                title: `Reveal in ${appProvider.systemUtils.isMac ?
-                    'Finder' : 'File Explorer'}`,
-                onClick: () => {
-                    openExplorer(dirPath);
-                },
-            },
-        ]);
-    };
     return (
-        <div className='path-selector w-100'
-            onContextMenu={onContextMenu}>
+        <div className='path-selector w-100'>
             <div className='d-flex path-previewer pointer'
                 onClick={() => {
                     setShowing(!showing);
@@ -64,14 +40,6 @@ export default function PathSelector({
     );
 }
 
-// TODO: check direction rtl error with /*
-const cleanPath = (path: string) => {
-    if (path && path[0] === '/') {
-        path = path.substring(1);
-    }
-    return path;
-};
-
 function RenderTitle({ dirSource }: { dirSource: DirSource }) {
     const [dirPath, setDirPath] = React.useState(dirSource.dirPath);
     useDSEvents(['path'], dirSource, () => {
@@ -82,10 +50,7 @@ function RenderTitle({ dirSource }: { dirSource: DirSource }) {
     }
     return (
         <>
-            <div className='ellipsis-left border-white-round px-1 flex-fill'
-                title={cleanPath(dirPath)}>
-                {cleanPath(dirPath)}
-            </div>
+            {pathPreviewer(dirPath)}
             <div className='px-2'
                 onClick={(event) => {
                     event.stopPropagation();
