+++ src/slide-presenting/items/SlidePreviewerFooter.tsx
@@ -1,3 +1,4 @@
+import { pathPreviewer } from '../../others/PathPreviewer';
 import Slide from '../../slide-list/Slide';
 import {
     DEFAULT_THUMBNAIL_SIZE,
@@ -7,32 +8,37 @@ import {
 } from '../../slide-list/slideHelpers';
 
 export default function SlidePreviewerFooter({
-    thumbnailSize, setThumbnailSize,
+    thumbnailSize, setThumbnailSize, slide,
 }: {
     thumbnailSize: number,
     setThumbnailSize: (size: number) => void,
+    slide: Slide,
 }) {
     const currentScale = (thumbnailSize / DEFAULT_THUMBNAIL_SIZE);
     return (
-        <div className='card-footer'>
-            <div className='d-flex justify-content-end h-100'>
-                <div className='size d-flex'>
-                    <label className='form-label'>
-                        Size:{currentScale.toFixed(1)}
-                    </label>
-                    <input type='range' className='form-range'
-                        min={MIN_THUMBNAIL_SCALE} max={MAX_THUMBNAIL_SCALE}
-                        step={THUMBNAIL_SCALE_STEP}
-                        value={currentScale.toFixed(1)}
-                        onChange={(event) => {
-                            setThumbnailSize((+event.target.value) *
-                                DEFAULT_THUMBNAIL_SIZE);
-                        }}
-                        onWheel={(event) => {
-                            const newScale = Slide.toScaleThumbSize(
-                                event.deltaY > 0, currentScale);
-                            setThumbnailSize(newScale * DEFAULT_THUMBNAIL_SIZE);
-                        }} />
+        <div className='card-footer w-100'>
+            <div className='d-flex justify-content-end w-100 h-100'>
+                <div className='size d-flex w-100 overflow-hidden'>
+                    {pathPreviewer(slide.fileSource.filePath)}
+                    <div className='form form-inline d-flex flex-row-reverse'
+                    style={{ minWidth: '100px' }}>
+                        <label className='form-label'>
+                            Size:{currentScale.toFixed(1)}
+                        </label>
+                        <input type='range' className='form-range'
+                            min={MIN_THUMBNAIL_SCALE} max={MAX_THUMBNAIL_SCALE}
+                            step={THUMBNAIL_SCALE_STEP}
+                            value={currentScale.toFixed(1)}
+                            onChange={(event) => {
+                                setThumbnailSize((+event.target.value) *
+                                    DEFAULT_THUMBNAIL_SIZE);
+                            }}
+                            onWheel={(event) => {
+                                const newScale = Slide.toScaleThumbSize(
+                                    event.deltaY > 0, currentScale);
+                                setThumbnailSize(newScale * DEFAULT_THUMBNAIL_SIZE);
+                            }} />
+                    </div>
                 </div>
             </div>
         </div>
